New: length-list-<, length-<
These are functions for testing whether a list or
sequence is shorter than a given integer. This is cheaper
than calculating the length of lists, which is in
some cases impossible if they are infinite.

A length-str-< function already exists, useful
with lazy strings.

length-< uses length-list-< or length-str-<
as appropriate

* lib.[ch] (length_list_lt, length_lt): New functions.

* eval.c (eval_init): length-list-< and length-<
intrinsics registered.

* tests/012/seq.tl: New tests.

* txr.1: Documented.