glob: suppress consecutive duplicates; fix memleak.
* glob.c (glob_wrap): When converting the glob array to the
returned list, suppress consecutive duplicates. This has to be
done separately for each call to glob or super_glob, so we now
interleave the production of the output list with the glob
calls.  It has to be done separately because there can be
duplicates between different patterns. E.g. if (glob "?")
matches one path then (glob '("?" "?")) must return two copies
of it. Furthermore, the brace expansion implementation in
glob* produces multiple glob calls and appends their results.
Duplicates inside a single super_glob call result when there
are multiple ** (double star) patterns present, which are
matched by the same path in more than one way. If the results
are sorted, then the duplicates appear consecutively and we
will squash them. Also, a memory leak is fixed here: we
must free(pat_u8) unconditionally, before testing for the
early exit situation.