listener: fix several bugs in auto compound expr mode.
* parser.c (repl): The first bug is that we are not correctly
checking the special variable: auto_parens holds the binding.
Thus TXR was behaving as if they feature is always enabled.
The second bug is that forms might not be a list; it could
be the colon symbol, so we cannot evaluate cdr(forms).
The third bug is that we don't want to create (progn . :)
when forms is the : symbol. These two bugs are reproduced
by turning on the mode and evaluating (1 2 3 . 4 5 6),
a bad form.