rperm: support general sequences.
* combi.c (rperm_init): Take one more parameter.
Allocate the state vector to three elements and
put the extra value there.
(rperm_list, rperm_vec, rperm_str): Pass nil
for the extra value to rperm_init; these do not
use it.
(rperm_seq_gen_fun, rperm_seq): New functions.
These use the extra value to store the original
seq in the state, so that the elements of the
output sequence can be converted to the same type
of sequence as the input.
(rperm): Replace error throw in default case with
call to rperm_seq.