read-objects: clean up working, but odd code.
This was prompted by a misleading indentation warning.

* parser.c (read_objects_common): The indented statment
"return error_return_val;" belongs with the if. This works
anyway because if lisp_parse_impl returned unique_s, it
means that either pi->syntax_tree is nao, or pi->errors
is positive. So, let's take advantage of that and don't
bother checking for pi->syntax_tree == nao. If unique_s
was returned and there are no errors, we hit EOF and
so break out of the loop.