tests: fix tests/007/except-4.txr for Solaris, BSD.
* tests/007/except-4.txr: The portable way to get a shell
command that exits with a signal is to execute kill -KILL $$.
If we use a signal that the shell catch like SIGTERM or
SIGINT, we get nonportable behaviors. Some shells seem to
catch the signal and then raise it again so they terminate
with that signal. Some shells terminate normally, but create
an exit status by OR-ing 0x80 with the caught signal.
Let's use kill -KILL here and drop the tests for BSD and
Solaris.

* tests/007/except-3.txr: Fix the kill command here also.
While this test wasn't failing on those platforms, it succeeds
vacuously, since the exception being ignored by :nothrow
is not actually thrown.