eval: potential gc problem in binding.
* eval.c (reparent_env): This function is used in bindings
helper, when special variables are involved. It makes a new
environment the parent of an existing one, just by assigning
the pointer. This is wrong under generational GC if it makes
a mature object point to a new object. We fix it with the set
macro. I've not seen a crash because of this; I caught it by
inspection.