sub: don't produce an iterator.
Having the sub function yield an iterator in some cases
is a defective requirement, causing problems like this:

  1> (partition 1..10 '(2 3))
  ((1 2) (3) #<seq-iter: a24c380>)

With fix:

  1> (partition 1..10 '(2 3))
  ((1 2) (3) (4 5 6 7 8 9))

* lib.c (sub_iter): When the interval is open and we are
operating on a sequence via iter-begin, do not return an
iterator. Convert it to a lazy list. Not subjecting this to -C
compat flag; I can't imagine anyone writing code to depend on
this, rather than stepping around it as a bugx.

* txr.1: Documentation updated.