zip: more permissive implementation.
zip and transpose should allow non-character data
when the leftmost column is a string, falling back
on making lists, like seq_build.

We can't use seq_build as-is because of the special
semantics of transpose/zip with regard to strings.
We introduce a "strcat" variant of seq_build
for this purpose.

* lib.c (seq_build_strcat_add): New static function.
(sb_strcat_ops): New static structure like sb_str_ops,
but with seq_build_strcat_add as the add operation,
which allows string arguments to be appended to the
string rather than switching to a list.
(seq_build_strcat_init): New function.

* lib.h (seq_build_strcat_init): Declared.

* eval.c (zip_strcat): New static function; uses
seq_build_strcat_init.
(zipv): Only recognize strings specially; all else goes
through the existing default case.
Strings use zip_strcat.

* tests/012/seq.tl: New test case.

* txr.1: Describe special semantics of zip/tranpose;
previously only documented in one example.
Clarify that the rows are only sequences of the
same kind as the leftmost column if possible,
otherwise lists.  Remove text which says that it's an error
for the other columns to contain non-string, non-character
objects if the leftmost column is a string.