keep-if, remove-if, keep-keys-if: mapfun argument.
We introduce a mapfun argument to these functions so that they
can additionally transform the accumulated values.

The keep-keys-if function is now implemented through the same
helper function as keep-if but with the mapfun argument
defaulting to a copy of the keyfun argument.

* eval.c (eval_init): Update registrations of remove-if,
keep-if and keep-keys-if to new arities of C functions.

* lib.c (rem_if_impl): Implement new optional mapfun
parameter.
(remove_if, keep_if): Add mapfun parameter.
(keep_keys_if): Implement via rem_if_impl, and add
mapfun argument. We do the defaulting of keyfun here,
so that we can then use that argument's value to default
mapfun.

* lib.h (remove_if, keep_if, keep_keys_if): Declarations
updated.

* tests/012/seq.tl: Couple of test cases exercising mapfun
argument of keep-if and remove-if.

* txr.1: Documented.