read-until-match: fix regression.
Commit 9aa751c8a4f845ef2d2bba091c81ffeded941afd
broke things.

This fix affects the function read-until-match,
scan-until-match and count-until-match which share
implementation.

* regex.c (scan_until_common): In the REGM_MATCH_DONE
and REGM_MATCH cases, we must push the character onto
the local stack, before doing the match = stack
assignment. Otherwise, it's possible that the stack
is empty and so no match is recorded. The REGM_FAIL
case will then behave as if no match was found, consuming
a character and continuing.

* txr.1: Codify an existing behavior: only non-empty matches
for the regex are considered by read-until-match.

* tests/015/regex.tl: New file. I am amazed to discover
that we don't seem to have a test suite for regexes at all.
Putting the tests here which confirm this fix and provide
coverage for some edge cases in read-until-match.