regex: bugfix: not taking full advantage of REGM_MATCH_DONE.
* regex.c (nfa_has_transitions): The logical disjunction here
is wrong.  We would like to test whether a state has
transitions if it is not an epsilon state. The code which uses
this macro doesn't care about epsilon states, even though they
have transitions; those are squeezed out by transitively
closure. The wrong condition here makes the code think that
a NFA set has transitions when it does not, preventing the
result code REGM_MATCH_DONE to be produced which can spare
a character from being consumed from a stream.